+++ src/oms_sensemaking/core/exceptions.py
@@ -1,0 +1,11 @@
+"""Module for sensemaking exceptions"""
+
+
+class SensemakingError(Exception):
+    """Sensemaking Exception"""
+
+
+class TrackLengthError(SensemakingError):
+    """
+    Tracks must have at least two points
+    """